adding test to fixture to see if it works
