+++ test/Cli/Func.E2E.Tests/Commands/FuncStart/TestsWithFixtures/NodeV4Tests.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
+using System.Diagnostics;
 using System.Net;
 using System.Net.Sockets;
 using Azure.Functions.Cli.E2E.Tests.Fixtures;
@@ -11,6 +12,7 @@ using Azure.Functions.Cli.TestFramework.Helpers;
 using FluentAssertions;
 using Xunit;
 using Xunit.Abstractions;
+using static Microsoft.Azure.AppService.Proxy.Runtime.Trace;
 
 namespace Azure.Functions.Cli.E2E.Tests.Commands.FuncStart.TestsWithFixtures
 {
@@ -175,6 +177,56 @@ namespace Azure.Functions.Cli.E2E.Tests.Commands.FuncStart.TestsWithFixtures
             // Validate failure message
             result.Should().ExitWith(1);
             result.Should().HaveStdErrContaining("The runtime argument value provided, 'inproc8', is invalid. The provided value is only valid for the worker runtime 'dotnet'.");
+        }
+
+        [Fact]
+        public void Start_FunctionApp_WhichExceedsTimeout_ShouldKillProcess()
+        {
+            var port = ProcessHelper.GetAvailablePort();
+            var testName = nameof(Start_FunctionApp_WhichExceedsTimeout_ShouldKillProcess);
+
+            // Start the function app with a process handler that intentionally stalls
+            var funcStartCommand = new FuncStartCommand(_fixture.FuncPath, testName, _fixture.Log);
+            var stopwatch = new Stopwatch();
+
+            funcStartCommand.ProcessStartedHandler = async (process) =>
+            {
+                try
+                {
+                    stopwatch.Start();
+
+                    // Wait for the function host to start
+                    await ProcessHelper.WaitForFunctionHostToStart(process, port, funcStartCommand.FileWriter ?? throw new ArgumentNullException(nameof(funcStartCommand.FileWriter)));
+
+                    // Log that we're starting the intentional stall
+                    _fixture.Log.WriteLine("Process started successfully. Intentionally stalling for longer than the timeout period (2 minutes)...");
+                    funcStartCommand.FileWriter?.WriteLine("[STDOUT] Intentionally stalling process for longer than timeout period...");
+                    funcStartCommand.FileWriter?.Flush();
+
+                    // Stall for 3 minutes (longer than the 2-minute timeout)
+                    // This should trigger the timeout in Command.cs
+                    await Task.Delay(TimeSpan.FromMinutes(3));
+                }
+                catch (Exception ex)
+                {
+                    // Log any unexpected exceptions
+                    string unhandledException = $"Unexpected exception: {ex}";
+                    _fixture.Log.WriteLine(unhandledException);
+                    funcStartCommand.FileWriter?.WriteLine("[STDOUT] unhandledException");
+                    funcStartCommand.FileWriter?.Flush();
+                }
+            };
+
+            // Execute the command
+            var result = funcStartCommand
+                .WithWorkingDirectory(_fixture.WorkingDirectory)
+                .WithEnvironmentVariable(Common.Constants.FunctionsWorkerRuntime, "node")
+                .Execute(["--port", port.ToString()]);
+
+            // Verify that the process was killed and didn't run for the full 3 minutes
+            // We expect it to be killed after 2 minutes (120 seconds) with some buffer
+            stopwatch.Elapsed.TotalSeconds.Should().BeLessThan(180);
+            stopwatch.Elapsed.TotalSeconds.Should().BeGreaterThan(110);
         }
     }
 }
